backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
install==1.3.5
mysql-connector-python==8.1.0
protobuf==4.21.12
pycodestyle==2.10.0
Pygments==2.16.1
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
